New admin release! Also a new Resolve Link grima!
This release improves the functionality and documentation of the
administrative grimas for grima user management. The new Resolve Link
grima can be used to resolve a link from Alma Preview or Primo Display.
This release also contains a few bugfixes and improvements for older
grimas. Dependencies are also updated to the current versions.